session_supervisor: carve registry.rs (pure move)
